backend/pyproject.toml
[tool.poetry]
name = "insurance-architecture-backend"
version = "0.1.0"
description = "LLM proxy with web interface"
authors = []
readme = "README.md"
packages = [{ include = "app" }, { include = "db" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.36" }
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
email-validator = "^2.2.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.17"
httpx = "^0.28.0"
aiofiles = "^24.1.0"
psycopg = {extras = ["binary"], version = "^3.3.3"}
argon2-cffi = "^25.1.0"
reportlab = "^4.4.4"
python-docx = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
init_db = "db.database:run_init_models"
run_server = "app.main:run_server"